refactor(config): extract source test modules
- Move config unit tests into sibling *_test.rs files
- Keep configuration modules focused on runtime resolution logic
